Split projects model
