+++ packages/models-library/src/models_library/__init__.py
@@ -1,5 +1,10 @@
 """ osparc's service models library
 
 """
+#
+# NOTE:
+#   - "examples" = [ ...] keyword and NOT "example". See https://json-schema.org/understanding-json-schema/reference/generic.html#annotations
+#
+
 
 __version__ = "0.1.0"
