+++ packages/models-library/src/models_library/projects_nodes.py
@@ -1,104 +1,22 @@
-from enum import Enum
+"""
+    Models Node as a central element in a project's pipeline
+"""
+
 from typing import Dict, List, Optional, Union
-from uuid import UUID
 
-from pydantic import AnyUrl, BaseModel, Extra, Field, HttpUrl, constr, validator
+from pydantic import BaseModel, Extra, Field, HttpUrl, constr, validator
 
+from .projects_access import AccessEnum
+from .projects_nodes_io import (
+    DatCoreFileLink,
+    DownloadLink,
+    NodeID,
+    PortLink,
+    SimCoreFileLink,
+)
+from .projects_nodes_ui import Position
+from .projects_state import RunningState
 from .services import KEY_RE, PROPERTY_KEY_RE, VERSION_RE
-
-NodeID = UUID
-
-
-class RunningState(str, Enum):
-    UNKNOWN = "UNKNOWN"
-    PUBLISHED = "PUBLISHED"
-    NOT_STARTED = "NOT_STARTED"
-    PENDING = "PENDING"
-    STARTED = "STARTED"
-    RETRY = "RETRY"
-    SUCCESS = "SUCCESS"
-    FAILED = "FAILED"
-    ABORTED = "ABORTED"
-
-
-class Position(BaseModel):
-    x: int = Field(..., description="The x position", example=["12"])
-    y: int = Field(..., description="The y position", example=["15"])
-
-    class Config:
-        extra = Extra.forbid
-
-
-class AccessEnum(str, Enum):
-    ReadAndWrite = "ReadAndWrite"
-    Invisible = "Invisible"
-    ReadOnly = "ReadOnly"
-
-
-class PortLink(BaseModel):
-    node_uuid: NodeID = Field(
-        ...,
-        description="The node to get the port output from",
-        example=["da5068e0-8a8d-4fb9-9516-56e5ddaef15b"],
-        alias="nodeUuid",
-    )
-    output: str = Field(
-        ...,
-        description="The port key in the node given by nodeUuid",
-        regex=PROPERTY_KEY_RE,
-        example=["out_2"],
-    )
-
-    class Config:
-        extra = Extra.forbid
-
-
-class DownloadLink(BaseModel):
-    download_link: AnyUrl = Field(..., alias="downloadLink")
-    label: Optional[str]
-
-    class Config:
-        extra = Extra.forbid
-
-
-class BaseFileLink(BaseModel):
-    store: Union[str, int] = Field(
-        ...,
-        description="The store identifier, '0' or 0 for simcore S3, '1' or 1 for datcore",
-        example=["0", 1],
-    )
-    path: str = Field(
-        ...,
-        description="The path to the file in the storage provider domain",
-        example=[
-            "N:package:b05739ef-260c-4038-b47d-0240d04b0599",
-            "94453a6a-c8d4-52b3-a22d-ccbf81f8d636/d4442ca4-23fd-5b6b-ba6d-0b75f711c109/y_1D.txt",
-        ],
-    )
-
-    class Config:
-        extra = Extra.forbid
-
-
-class SimCoreFileLink(BaseFileLink):
-    pass
-
-
-class DatCoreFileLink(BaseFileLink):
-    dataset: str = Field(
-        ...,
-        description="Unique identifier to access the dataset on datcore (REQUIRED for datcore)",
-        example=["N:dataset:f9f5ac51-33ea-4861-8e08-5b4faf655041"],
-    )
-    label: str = Field(
-        ...,
-        description="The real file name (REQUIRED for datcore)",
-        example=["MyFile.txt"],
-    )
-
-    class Config:
-        extra = Extra.forbid
-
 
 InputTypes = Union[
     int, bool, str, float, PortLink, SimCoreFileLink, DatCoreFileLink, DownloadLink
@@ -142,6 +60,8 @@ class Node(BaseModel):
         example=["https://placeimg.com/171/96/tech/grayscale/?0.jpg"],
     )
 
+    # INPUT PORTS ---
+
     inputs: Optional[Inputs] = Field(
         default_factory=dict, description="values of input properties"
     )
@@ -155,6 +75,7 @@ class Node(BaseModel):
         alias="inputNodes",
     )
 
+    # OUTPUT PORTS ---
     outputs: Optional[Outputs] = Field(
         default_factory=dict, description="values of output properties"
     )
@@ -168,11 +89,9 @@ class Node(BaseModel):
 
     parent: Optional[NodeID] = Field(
         None,
-        description="Parent's (group-nodes') node ID s.",
+        description="Parent's (group-nodes') node ID s. Used to group",
         example=["nodeUUid1", "nodeUuid2"],
     )
-
-    position: Optional[Position] = Field(None, deprecated=True)
 
     state: Optional[RunningState] = Field(
         RunningState.NOT_STARTED,
@@ -180,8 +99,8 @@ class Node(BaseModel):
         example=["RUNNING", "FAILED"],
     )
 
-
-
+    # NOTE: use projects_ui.py
+    position: Optional[Position] = Field(None, deprecated=True)
 
     @validator("thumbnail", pre=True)
     @classmethod
