+++ packages/models-library/src/models_library/projects_state.py
@@ -1,7 +1,13 @@
-from typing import Optional
-from pydantic import BaseModel, Field, PositiveInt, Extra
+"""
+    Models both project and node states
+"""
 
 from enum import Enum
+from typing import Optional
+
+from pydantic import BaseModel, Extra, Field
+
+from .projects_access import Owner
 
 
 class RunningState(str, Enum):
@@ -12,21 +18,8 @@ class RunningState(str, Enum):
     STARTED = "STARTED"
     RETRY = "RETRY"
     SUCCESS = "SUCCESS"
-    FAILURE = "FAILURE"
+    FAILED = "FAILED"
     ABORTED = "ABORTED"
-
-
-class Owner(BaseModel):
-    user_id: PositiveInt = Field(
-        ...,
-        description="Owner's identifier when registered in the user's database table",
-        examples=[2],
-    )
-    first_name: str = Field(..., description="Owner first name", examples=["John"])
-    last_name: str = Field(..., description="Owner last name", examples=["Smith"])
-
-    class Config:
-        extra = Extra.forbid
 
 
 class ProjectLocked(BaseModel):
