+++ packages/models-library/src/models_library/projects_ui.py
@@ -1,18 +1,13 @@
+"""
+    Models Front-end UI
+"""
+
 from typing import Dict, Optional
 
-from pydantic import BaseModel, Extra, Field, constr
-from .basic_regex import UUID_RE
-from .project_nodes import Node, Position
-NodeID = constr(regex=UUID_RE)
+from pydantic import BaseModel, Extra, Field
 
-# Pydantic does not support exporting a jsonschema with Dict keys being something else than a str
-# this is a regex for having uuids of type: 8-4-4-4-12 digits
-_NodeIDForDict = constr(
-    regex=r"^[0-9a-fA-F]{8}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{12}$"
-)
-Workbench = Dict[_NodeIDForDict, Node]
-
-
+from .projects_nodes_io import NodeID, NodeID_AsDictKey
+from .projects_nodes_ui import Position
 
 
 class WorkbenchUI(BaseModel):
@@ -29,10 +24,9 @@ class Slideshow(BaseModel):
         extra = Extra.forbid
 
 
-
 class StudyUI(BaseModel):
-    workbench: Optional[Dict[_NodeIDForDict, WorkbenchUI]] = Field(None)
-    slideshow: Optional[Dict[_NodeIDForDict, Slideshow]] = Field(None)
+    workbench: Optional[Dict[NodeID_AsDictKey, WorkbenchUI]] = Field(None)
+    slideshow: Optional[Dict[NodeID_AsDictKey, Slideshow]] = Field(None)
     current_node_id: Optional[NodeID] = Field(alias="currentNodeId")
 
     class Config:
